+++ src/components/content/hooks/use-toggle-dismiss.ts
@@ -2,7 +2,6 @@ import { useCallback, useState } from "react";
 import { useCTW } from "@/components/core/providers/use-ctw";
 import { toggleDismiss } from "@/fhir/basic";
 import { FHIRModel } from "@/fhir/models/fhir-model";
-import { queryClient } from "@/utils/request";
 
 interface UseToggleDismissResult {
   /**
@@ -18,10 +17,8 @@ interface UseToggleDismissResult {
 
 /**
  * This hook is toggles the dismiss status for the specified FHIR model.
- *
- * @param queriesToInvalidate  Queries to refetch
  */
-export function useToggleDismiss(...queriesToInvalidate: string[]): UseToggleDismissResult {
+export function useToggleDismiss(): UseToggleDismissResult {
   const { getRequestContext } = useCTW();
 
   const [isLoading, setIsLoading] = useState<boolean>(false);
@@ -29,11 +26,6 @@ export function useToggleDismiss(...queriesToInvalidate: string[]): UseToggleDis
   const handleToggleDismiss = useCallback(async (model: FHIRModel<fhir4.Resource>) => {
     setIsLoading(true);
     await toggleDismiss(model, await getRequestContext());
-    await Promise.all(
-      queriesToInvalidate.map((queryToInvalidate) =>
-        queryClient.invalidateQueries([queryToInvalidate])
-      )
-    );
     setIsLoading(false);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);