+++ src/components/content/unread-records-notification.tsx
@@ -66,7 +66,7 @@ export function useUnreadRecordsByResource(resources = defaultZAPTabs): Map<ZAPT
   }
 
   for (const resource of resources) {
-    let data: FHIRModel<fhir4.Resource>[];
+    let data: FHIRModel<fhir4.Resource>[] | undefined;
     switch (resource) {
       case "allergies":
         data = allergiesQuery.data;
@@ -93,7 +93,7 @@ export function useUnreadRecordsByResource(resources = defaultZAPTabs): Map<ZAPT
         data = [];
     }
     if (
-      data.some(
+      data?.some(
         (record) => !record.isDismissed && !record.isRead && !record.ownedByBuilder(userBuilderId)
       )
     ) {