+++ src/fhir/diagnostic-report.ts
@@ -1,5 +1,4 @@
 import { useEffect, useState } from "react";
-import { useIncludeBasics } from "./basic";
 import { LOINC_ANALYTES } from "./models/observation";
 import { usePatientObservations } from "./observations";
 import { CTWRequestContext } from "@/components/core/providers/ctw-context";
@@ -21,7 +20,7 @@ export function usePatientDiagnosticReports(
   includeObservations = false,
   enabled = true
 ) {
-  const query = useQueryWithPatient(
+  return useQueryWithPatient(
     QUERY_KEY_PATIENT_DIAGNOSTIC_REPORTS,
     [limit, includeObservations], // Only use the IDs in our key (fixes issue with ciruclar references).
     withTimerMetric(
@@ -30,8 +29,6 @@ export function usePatientDiagnosticReports(
     ),
     enabled
   );
-
-  return useIncludeBasics(query);
 }
 
 // Gets diagnostic reports for the patient with trending data for each observation in the diagnostic report.
@@ -63,7 +60,7 @@ export function usePatientDiagnosticReportsWithTrendData(limit = MAX_OBJECTS_PER
         observation.diagnosticReport = new DiagnosticReportModel(
           diagnosticReport.resource,
           undefined,
-          undefined,
+          diagnosticReport.basics,
           observations
         );
       }
@@ -76,7 +73,7 @@ export function usePatientDiagnosticReportsWithTrendData(limit = MAX_OBJECTS_PER
           new DiagnosticReportModel(
             dr.resource,
             undefined,
-            undefined,
+            dr.basics,
             observationsWithDiagnosticReportBackLink
           )
       )
@@ -108,7 +105,7 @@ function getDiagnosticReports(limit: number, includeObservations: boolean) {
         Telemetry.countMetric(`req.count.diagnostic_reports.none`, 1);
       }
       const result = data.DiagnosticReportConnection.edges.map(
-        (x) => new DiagnosticReportModel(x.node)
+        (x) => new DiagnosticReportModel(x.node, undefined, x.node.BasicList)
       );
 
       Telemetry.histogramMetric(`req.count.diagnostic_reports`, result.length);
@@ -166,7 +163,9 @@ export async function fetchDiagnosticReportsFromFQSById(
         },
       }
     );
-    return data.DiagnosticReportConnection.edges.map((x) => new DiagnosticReportModel(x.node));
+    return data.DiagnosticReportConnection.edges.map(
+      (x) => new DiagnosticReportModel(x.node, undefined, x.node.BasicList)
+    );
   } catch (e) {
     throw new Error(`Failed fetching document information for patient: ${e}`);
   }