+++ src/fhir/document.ts
@@ -1,5 +1,4 @@
 import { useEffect, useState } from "react";
-import { useIncludeBasics } from "./basic";
 import { DocumentModel } from "./models/document";
 import { PatientModel, useQueryWithPatient } from "..";
 import { applyDocumentFilters } from "@/components/content/document/helpers/filters";
@@ -16,7 +15,7 @@ export function usePatientTopLevelDocuments(limit = MAX_OBJECTS_PER_REQUEST, ena
 
   useEffect(() => {
     const filteredDocuments = orderBy(
-      applyDocumentFilters(data),
+      applyDocumentFilters(data ?? []),
       "resource.content[0].attachment.creation",
       ["desc"]
     ) as DocumentModel[];
@@ -38,14 +37,12 @@ export function usePatientTopLevelDocuments(limit = MAX_OBJECTS_PER_REQUEST, ena
 }
 
 export function usePatientDocuments(limit = MAX_OBJECTS_PER_REQUEST, enabled = true) {
-  const patientDocumentsQuery = useQueryWithPatient(
+  return useQueryWithPatient(
     QUERY_KEY_PATIENT_DOCUMENTS,
     [limit],
     withTimerMetric(getDocumentsFromFQS(limit), "req.timing.documents"),
     enabled
   );
-
-  return useIncludeBasics(patientDocumentsQuery);
 }
 
 function getDocumentsFromFQS(limit: number) {
@@ -64,7 +61,9 @@ function getDocumentsFromFQS(limit: number) {
           },
         }
       );
-      return data.DocumentReferenceConnection.edges.map((x) => new DocumentModel(x.node));
+      return data.DocumentReferenceConnection.edges.map(
+        (x) => new DocumentModel(x.node, undefined, x.node.BasicList)
+      );
     } catch (e) {
       throw new Error(`Failed fetching document information for patient: ${e}`);
     }