+++ src/services/fqs/queries/documents.ts
@@ -1,12 +1,16 @@
 import { DocumentReference } from "fhir/r4";
 import { gql } from "graphql-request";
-import { fragmentCoding, fragmentOrganization } from "./fragments";
+import { fragmentBasic, fragmentCoding, fragmentOrganization } from "./fragments";
 import { GraphqlConnectionNode, GraphqlPageInfo } from "../client";
 
 export interface DocumentReferenceConnection {
   pageInfo: GraphqlPageInfo;
-  edges: GraphqlConnectionNode<DocumentReference>[];
+  edges: GraphqlConnectionNode<DocumentReferenceWithBasics>[];
 }
+
+export type DocumentReferenceWithBasics = DocumentReference & {
+  BasicList: fhir4.Basic[];
+};
 
 export interface DocumentReferenceGraphqlResponse {
   DocumentReferenceConnection: DocumentReferenceConnection;
@@ -15,6 +19,7 @@ export interface DocumentReferenceGraphqlResponse {
 export const documentsQuery = gql`
   ${fragmentCoding}
   ${fragmentOrganization}
+  ${fragmentBasic}
   query Documents(
     $upid: ID!
     $cursor: String!
@@ -36,6 +41,9 @@ export const documentsQuery = gql`
         node {
           id
           resourceType
+          BasicList(_reference: "subject") {
+            ...Basic
+          }
           meta {
             extension {
               url